requirements.txt
bidict==0.22.1
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
dnspython==2.3.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.2
pycparser==2.21
pymongo==4.3.3
python-dotenv==0.21.1
python-engineio==4.3.4
python-socketio==5.7.2
six==1.16.0
Werkzeug==2.2.2
